Refactor gradient initialization into modules
